requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
camelot-py==0.10.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
distro==1.6.0
Django==3.2.6
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
et-xmlfile==1.1.0
ghostscript==0.7
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
mysqlclient @ file:///D:/projects/rencana_kerja/mysqlclient-1.4.6-cp38-cp38-win32.whl
numpy==1.21.1
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.3.1
pdfminer.six==20201018
pdftopng==0.2.3
Pillow==8.3.1
promise==2.3
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyTelegramBotAPI==3.8.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
tabula-py==2.2.0
tabulate==0.8.9
text-unidecode==1.3
urllib3==1.26.6
